fixes to param island shader

+++ include/polyscope/parameterization_quantity.h
@@ -32,7 +32,7 @@ public:
   // Wrapper around the actual buffers of data stored in the class.
   // Interaction with the data (updating it on CPU or GPU side, accessing it, etc) happens through this wrapper.
   render::ManagedBuffer<glm::vec2> coords;
-  render::ManagedBuffer<int32_t> islandLabels; // only optionally populated
+  render::ManagedBuffer<float> islandLabels; // only optionally populated. should be integers.
 
   const ParamCoordsType coordsType;
 
@@ -72,7 +72,7 @@ public:
 protected:
   // Raw storage for the data. You should only interact with this via the managed buffer above
   std::vector<glm::vec2> coordsData;
-  std::vector<int32_t> islandLabelsData;
+  std::vector<float> islandLabelsData;
   bool islandLabelsPopulated = false;
 
   // === Visualization parameters
