+++ include/polyscope/surface_parameterization_quantity.h
@@ -90,7 +90,7 @@ protected:
 template <typename V>
 void SurfaceParameterizationQuantity::setIslandLabels(const V& newIslandLabels) {
   validateSize(newIslandLabels, this->nFaces(), "scalar quantity " + quantity.name);
-  islandLabels.data = standardizeArray<int, V>(newIslandLabels);
+  islandLabels.data = standardizeArray<float, V>(newIslandLabels);
   islandLabels.markHostBufferUpdated();
   islandLabelsPopulated = true;
 }
